Refactor route handling classes into BaseExtensionRestHandler
