+++ src/main/java/org/opensearch/sdk/sample/helloworld/rest/RestHelloAction.java
@@ -18,8 +18,6 @@ import org.opensearch.extensions.rest.ExtensionRestResponse;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.sdk.rest.BaseExtensionRestHandler;
 import org.opensearch.sdk.rest.ExtensionRestHandler;
-import org.opensearch.sdk.rest.RouteHandler;
-
 import java.io.IOException;
 import java.net.URLDecoder;
 import java.nio.ByteBuffer;
